requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
dlib==19.23.1
docker-pycreds==0.4.0
filelock==3.6.0
fonttools==4.31.2
gdown==4.4.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.11.2
kiwisolver==1.4.2
matplotlib==3.5.1
ninja==1.10.2.3
numpy==1.21.5
opencv-python==4.5.5.62
packaging==21.3
pathtools==0.1.2
Pillow==9.0.1
piq==0.6.0
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
scipy==1.7.3
sentry-sdk==1.5.6
setproctitle==1.2.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2
termcolor==1.1.0
torch==1.7.1
torch-fidelity==0.3.0
torchvision==0.8.2
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
wandb==0.12.11
yaspin==2.1.0
zipp==3.7.0
